app.py
from fastapi import FastAPI, File, UploadFile, HTTPException, Depends
from fastapi.security import HTTPBearer, HTTPAuthorizationCredentials
from typing import List, Dict, Any
import os
from dotenv import load_dotenv
import uvicorn

from models.schema import (
    QuestionRequest,
    QuestionResponse,
    ComplexQueryRequest,
    ComplexQueryResponse
)

# Load environment variables
load_dotenv()

# Initialize FastAPI application
app = FastAPI(
    title="Document QA System",
    description="AI-powered document question answering system",
    version="1.0.0"
)

# API Endpoints
@app.get("/")
async def root():
    """Health check endpoint"""
    return {"status": "ok", "message": "Document QA System API"}

@app.post("/documents/upload")
async def upload_document(
    file: UploadFile = File(...),
):
    """
    Upload and process a document file (PDF, TXT, etc.)
    """
    # Implement document upload and processing
    return {"status": "not_implemented"}

@app.get("/documents")
async def get_documents():
    """
    Get all uploaded documents
    """
    # Implement document listing
    return {"status": "not_implemented"}

@app.post("/questions/ask", response_model=QuestionResponse)
async def ask_question(
    request: QuestionRequest,
):
    """
    Ask a question about uploaded documents
    """
    # Implement question answering
    return {"status": "not_implemented"}

@app.post("/agents/complex-query", response_model=ComplexQueryResponse)
async def complex_query(
    request: ComplexQueryRequest,
):
    """
    Process a complex query using multiple agents
    """
    # Implement complex query processing
    return {"status": "not_implemented"}

if __name__ == "__main__":
    uvicorn.run("app:app", host="0.0.0.0", port=8000, reload=True)